+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/UpdateTypes.kt
@@ -9,6 +9,7 @@ const val UPDATE_INLINE_QUERY = "inline_query"
 const val UPDATE_CALLBACK_QUERY = "callback_query"
 const val UPDATE_SHIPPING_QUERY = "shipping_query"
 const val UPDATE_PRE_CHECKOUT_QUERY = "pre_checkout_query"
+const val UPDATE_POLL = "poll"
 
 val ALL_UPDATES_LIST = listOf(
     UPDATE_MESSAGE,
@@ -19,5 +20,6 @@ val ALL_UPDATES_LIST = listOf(
     UPDATE_INLINE_QUERY,
     UPDATE_CALLBACK_QUERY,
     UPDATE_SHIPPING_QUERY,
-    UPDATE_PRE_CHECKOUT_QUERY
+    UPDATE_PRE_CHECKOUT_QUERY,
+    UPDATE_POLL
 )