+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/extensions/UpdatesPolling.kt
@@ -34,6 +34,7 @@ suspend fun RequestsExecutor.startGettingOfUpdates(
     callbackQueryCallback: UpdateReceiver<CallbackQueryUpdate>? = null,
     shippingQueryCallback: UpdateReceiver<ShippingQueryUpdate>? = null,
     preCheckoutQueryCallback: UpdateReceiver<PreCheckoutQueryUpdate>? = null,
+    pollCallback: UpdateReceiver<PollUpdate>? = null,
     requestsDelayMillis: Long = 1000,
     scope: CoroutineScope = GlobalScope
 ): Job {
@@ -50,7 +51,8 @@ suspend fun RequestsExecutor.startGettingOfUpdates(
         inlineQueryCallback,
         callbackQueryCallback,
         shippingQueryCallback,
-        preCheckoutQueryCallback
+        preCheckoutQueryCallback,
+        pollCallback
     )
     return startGettingOfUpdates(
         requestsDelayMillis,
@@ -71,6 +73,7 @@ suspend fun RequestsExecutor.startGettingOfUpdates(
     callbackQueryCallback: UpdateReceiver<CallbackQueryUpdate>? = null,
     shippingQueryCallback: UpdateReceiver<ShippingQueryUpdate>? = null,
     preCheckoutQueryCallback: UpdateReceiver<PreCheckoutQueryUpdate>? = null,
+    pollCallback: UpdateReceiver<PollUpdate>? = null,
     requestsDelayMillis: Long = 1000,
     scope: CoroutineScope = GlobalScope
 ): Job = startGettingOfUpdates(
@@ -87,6 +90,7 @@ suspend fun RequestsExecutor.startGettingOfUpdates(
     callbackQueryCallback = callbackQueryCallback,
     shippingQueryCallback = shippingQueryCallback,
     preCheckoutQueryCallback = preCheckoutQueryCallback,
+    pollCallback = pollCallback,
     requestsDelayMillis = requestsDelayMillis,
     scope = scope
 )